Move all logic in profile360 to profile
