+++ src/applications/personalization/profile/components/AccountSecurity.jsx
@@ -4,7 +4,7 @@ import DowntimeNotification, {
 } from 'platform/monitoring/DowntimeNotification';
 import { focusElement } from 'platform/utilities/ui';
 
-import { handleDowntimeForSection } from 'applications/personalization/profile360/components/DowntimeBanner';
+import { handleDowntimeForSection } from 'applications/personalization/profile/components/DowntimeBanner';
 
 import AccountSecurityContent from './AccountSecurityContent';
