+++ src/applications/personalization/profile/components/personal-information/PersonalInformation.jsx
@@ -9,9 +9,9 @@ import DowntimeNotification, {
 } from 'platform/monitoring/DowntimeNotification';
 import { focusElement } from 'platform/utilities/ui';
 
-import PaymentInformationBlocked from 'applications/personalization/profile360/components/PaymentInformationBlocked';
-import { handleDowntimeForSection } from 'applications/personalization/profile360/components/DowntimeBanner';
-import { directDepositIsBlocked } from 'applications/personalization/profile360/selectors';
+import PaymentInformationBlocked from 'applications/personalization/profile/components/PaymentInformationBlocked';
+import { handleDowntimeForSection } from 'applications/personalization/profile/components/DowntimeBanner';
+import { directDepositIsBlocked } from 'applications/personalization/profile/selectors';
 
 import PersonalInformationContent from './PersonalInformationContent';
