+++ src/applications/personalization/profile/manifest.json
@@ -1,0 +1,6 @@
+{
+  "appName": "VA Profile",
+  "entryFile": "./profile-entry.jsx",
+  "entryName": "profile",
+  "rootUrl": "/profile"
+}
