+++ src/applications/personalization/profile/sass/profile-360.scss
