+++ src/applications/personalization/profile/sass/user-profile.scss
