--- src/applications/personalization/profile360/constants/index.js
@@ -1,4 +1,0 @@
-export const ACCOUNT_TYPES_OPTIONS = {
-  checking: 'Checking',
-  savings: 'Savings',
-};
