--- src/applications/personalization/profile360/manifest.json
@@ -1,6 +1,0 @@
-{
-  "appName": "VA Profile",
-  "entryFile": "./profile-360-entry.jsx",
-  "entryName": "profile-360",
-  "rootUrl": "/profile"
-}
